feat: arg validators return an error
The ArgValidator now returns an error instead of a bool. This allows
for better `invalid args` error messages to the user.

Previously if a command expected one arg and instead recieved 2, the
user simply had to guess at what `invalid args` would mean, but in this
case, the `ArgsExact(1)` will return an error `expected 1 arg, got 2`
which is much more meaningful.